services/poller/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy[asyncio]
asyncpg
apscheduler
python-dotenv
pydantic
requests